+++ src/main/java/com/olx/boilerplate/infrastructure/logging/LoggingConfiguration.java
@@ -1,6 +1,6 @@
 package com.olx.boilerplate.infrastructure.logging;
 
-import com.olx.boilerplate.domain.port.AppLoggers;
+import com.olx.boilerplate.logging.AppLoggers;
 import jakarta.annotation.PostConstruct;
 import org.springframework.context.annotation.Configuration;